TEIIDSB-206 updating dependencies and style changes to the test class
